+++ src/pages/additional-services/99apartment-services/index.tsx
@@ -43,7 +43,7 @@ export default function AptAdditionalServicesList() {
           <div className=" w-full flex justify-between">
             <h2 className="text-xl font-semibold"> All Service List</h2>
             <div className=" flex items-center flex-col md:flex-row gap-4">
-              <ExportSelect id="customers" />
+              {/* <ExportSelect id="customers" /> */}
               <LoadingButton
                 clickHandler={() => setOpenNewRequest(true)}
                 isLoading={false}